Refactor some compat-handling utilities out of `dist.ts`
